small changes to the babylon-server
