+++ packages/tools/babylonServer/.gitignore
@@ -1,1 +1,2 @@
 declarations/
+create*.*
