+++ packages/tools/babylonServer/tsconfig.json
@@ -2,5 +2,8 @@
     "extends": "../../../tsconfig.json",
 
     "compilerOptions": {
+        "types": [
+            "declarations/**/*.d.ts"
+        ]
     },
 }
